refactor: extract buildStepOverrides and split model profiles per review
